+++ src/StateSmith/compiler/ExitPointValidator.cs
@@ -1,16 +1,24 @@
 ﻿namespace StateSmith.Compiling;
+using StateSmith.compiler;
+using System.Linq;
 
 public class ExitPointValidator
 {
-    public static void Validate(ExitPoint v)
+    public static void Validate(ExitPoint exitPoint)
     {
-        PseudoStateValidator.ValidateParentAndNoChildren(v);
+        PseudoStateValidator.ValidateParentAndNoChildren(exitPoint);
 
-        if (v.IncomingTransitions.Count == 0)
+        if (exitPoint.IncomingTransitions.Count == 0)
         {
-            throw new VertexValidationException(v, "An exit point must at least one incoming transition (for now).");
+            throw new VertexValidationException(exitPoint, "An exit point must at least one incoming transition (for now).");
         }
 
-        PseudoStateValidator.ValdiateExitingBehaviors(v);
+        PseudoStateValidator.ValidateExitingBehaviors(exitPoint);
+
+        var duplicateLabelCount = exitPoint.SiblingsOfMyType().Where(v => v.label == exitPoint.label).Count();
+        if (duplicateLabelCount > 0)
+        {
+            throw new VertexValidationException(exitPoint, $"Found {duplicateLabelCount} exit point(s) with duplicate label `{exitPoint.label}` within the same immediate parent.");
+        }
     }
 }