+++ src/StateSmithTest/ExitPointValidationTests.cs
@@ -1,4 +1,5 @@
 ﻿using StateSmith.Compiling;
+using Xunit;
 
 namespace StateSmithTest.PseudoStateTests
 {
@@ -18,6 +19,26 @@ namespace StateSmithTest.PseudoStateTests
         override protected ExitPoint CreateS2PseudoState()
         {
             return new ExitPoint("1");
+        }
+
+        [Fact]
+        public void SameNameOkIfNotInSameScope()
+        {
+            var duplicateNamed = new ExitPoint("1");
+            s2_1.AddChild(duplicateNamed);
+            duplicateNamed.AddTransitionTo(s2_1).guardCode = "true"; // to get around yEd concern
+            s2_1_1.AddTransitionTo(duplicateNamed);
+            ExpectValid();
+        }
+
+        [Fact]
+        public void ErrorIfDuplicateNameInParent()
+        {
+            var duplicateNamed = new ExitPoint("1");
+            s2.AddChild(duplicateNamed);
+            duplicateNamed.AddTransitionTo(s2_1);
+            s2_1.AddTransitionTo(duplicateNamed);
+            ExpectVertexValidationExceptionWildcard("*duplicate*");
         }
     }
 }